app/pyproject.toml
[tool.poetry]
name = "finance-analytics"
version = "0.2.0"
description = "Finance analytics platform: scraping, parsing, scoring, sentiment, and summary generation"
authors = ["Finance Analytics Team"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.0"
lxml = "^5.1.0"
python-dotenv = "^1.0.0"
psycopg = {extras = ["binary", "pool"], version = "^3.1.0"}
boto3 = "^1.34.0"
prefect = "^2.14.0"
playwright = "^1.41.0"
feedparser = "^6.0.10"
click = "^8.1.0"
tenacity = "^8.2.0"
rich = "^13.7.0"
pdfplumber = "^0.11.0"
transformers = "^4.37.0"
torch = "^2.2.0"
yfinance = "^0.2.36"
scikit-learn = "*"
lightgbm = "^4.1.0"



[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
black = "^24.1.0"
ruff = "^0.1.0"
mypy = "^1.8.0"

[tool.poetry.scripts]
finance = "src.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
